refactor(main): extract reading-report control surface and annotation UI
- lib/report_ui.lua: the reading-time report menu (enable, trigger mode,
  auto-association vs a manually picked target book, status), the target
  book picker, and reading statistics. The reporting state machine stays
  in lib/read_report.lua; this is only its control surface.
- lib/annotations_ui.lua: showing/hiding the underlines baked into cached
  EPUBs and turning a tap on one into a thought popup. It now owns the
  reader session generation, so a popup or highlight-clear scheduled for a
  previous book can no longer act on the current one, and the three
  subtleties that make this code fragile (stylesheet-only visibility
  applied from onReadSettings, the getLinkFromGes filter working around
  crengine ignoring pointer-events, and session-generation guards) are
  documented at the top of the module.

main.lua keeps thin onReadSettings / onReaderReady / onCloseDocument
handlers that delegate to the two controllers.

Bodies were moved verbatim; only call paths changed. i18n keys unchanged.

main.lua: 1270 -> 510 lines.
